backend/requirements.txt
fastapi
uvicorn
python-multipart
whisper
torch
google-genai
numpy
faiss-cpu
sentence-transformers
pymongo
